requirements.txt
Pillow>=10.0.0
requests>=2.31.0
cairosvg>=2.7.0
numpy>=1.24.0
realesrgan>=0.3.0
basicsr>=1.4.2
torch>=2.0.0
torchvision>=0.15.0
pywebview>=5.3
